finalize milticore stabilizer code
